requirements.txt
fastapi
uvicorn
torch
numpy
pillow
PyMuPDF
doclayout_yolo
huggingface_hub
python-multipart
boto3
requests
aiohttp